daemon.rs was 3,261 lines; DES-09 says 600, and nothing was checking
`docs/engineering/design.md` DES-09 is not a style preference — it names
the failure: every service in this repo that passed roughly 600 lines
also acquired a state bug a reviewer had to reconstruct the whole file to
find. The gate that enforces it, `check_architecture.py`, reads Python
only. The desktop crates were never measured, and `main.rs` reached
11,297 lines with nothing objecting.

daemon.rs is now eleven modules, split by use case as DES-10 asks rather
than by layer: dispatch, monitors, sharing_ops, config_ops,
agent_host_ops, stack_ops, reset_ops, supervisor, environment, the tests,
and a mod.rs holding construction, the client loop and the broadcast
plumbing. Largest is 512 lines. Every method moved verbatim; the 256
tests pass unchanged, which is the point of doing it this way.

The part worth reviewing is not the moves, it is what the moves nearly
broke. Four guards scanned `daemon.rs` by name — the doc-comment shape,
the no-proxy client rule, console-window suppression, and the one in the
shell that enumerates locald's redirect variables. A split silently
shrinks all four to whichever file keeps the old name.

So: locald's three now share one `DAEMON_SOURCES` list, and a new test
fails if a file appears under `src/daemon/` without being added to it.
The shell's guard reads the directory from disk instead — its own comment
says "a list somebody maintains is how the first three went missing", and
eleven hand-written entries would have been exactly that list.

And DES-09 now reaches these crates: `check_file_size.py` is a ratchet in
the shape of the Python one, baselined at today's twenty offenders, with
`make desktop-check` running it. A file over the limit may only shrink; a
file under it may not cross. Verified it bites by shrinking a baseline
entry by five lines.
